+++ app/src/main/java/org/endhungerdurham/pantries/ui/PantriesFragment.kt
@@ -1,10 +1,11 @@
-package org.endhungerdurham.pantries
+package org.endhungerdurham.pantries.ui
 
 import android.support.v4.app.Fragment
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import org.endhungerdurham.pantries.R
 
 class PantriesFragment : Fragment() {
