+++ app/src/main/res/layout/fragment_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ListFragment">
+    tools:context=".ui.ListFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/list"
